Add support for specifying CAA records
This allows specifying which CA authorities are allowed to issue
certificates for subdomains or disallow issuing certificates entirely.
This is disabled by default and is intended for situations where having
valid TLS certificates is not necessary and it's not desirable to have a
malicious IP (eg. phishing site) potentially appear more legitimate due
to the valid certificate.

This commit includes a couple minor fixes:

- _get_env_splitted() now calls split(..., 1) instead of split(..., 2),
  which actually produces up to 3 parts.
- The individual _write("END") calls in various qtype handlers were
  replaced by a single write_end() call at the end of run()'s loop.